chore: ignore part with canvas and window
